# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: House IoT API
#host: 172.24.1.1:10010
host: localhost:10010 #On MAC localhost
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /status:
    x-swagger-router-controller: device
    get:
      description: Returns device status
      operationId: getStatus
      tags:
        - Device
      parameters:
        - name: deviceName
          in: query
          description: The name of the device
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: putStatus
      description: Updates device status value
      tags:
        - Device
      parameters:
        - name: deviceName
          in: query
          description: The name of the device
          required: true
          type: string
        - name: newValue
          description: This is the new status value
          in: body
          required: true
          schema:
            $ref: "#/definitions/newValue"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /value:
    x-swagger-router-controller: device
    get:
      description: Returns device value
      operationId: getValue
      tags:
        - Device
      parameters:
        - name: deviceName
          in: query
          description: The name of the device
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: putValue
      description: Updates device value
      tags:
        - Device
      parameters:
        - name: deviceName
          in: query
          description: The name of the device
          required: true
          type: string
        - name: newValue
          description: This is the new value
          in: body
          required: true
          schema:
            $ref: "#/definitions/newValue"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /info:
    x-swagger-router-controller: device
    get:
      description: Returns all device info
      # used as the method name of the controller
      operationId: getInfo
      tags:
        - Device
      parameters:
        - name: deviceName
          in: query
          description: The name of the device
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Device"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  SuccessResponse:
    required:
      - value
    properties:
      value:
        type: integer
        description: This is the response value, status or value of device
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Device:
    properties:
      name:
        type: string
        description: This is the device name
      status:
        type: integer
        description: This is the device status (1 is active 0 is offline)
      value:
        type: integer
        description: This is the device value (1 is on 0 is off)
      pin:
        type: integer
        description: This is the pin of ESP8266 where the device is connect
      info:
        type: string
        description: This is the device description
  newValue:
    properties:
      value:
        type: integer
        description: This is the new device status or new device value